+++ packages/frontend/src/components/SimpleTable/index.tsx
@@ -3,8 +3,12 @@ import {
     AdditionalMetric,
     Field,
     friendlyName,
+    getAdditionalMetricLabel,
+    getItemLabel,
     getItemMap,
     getResultValues,
+    isAdditionalMetric,
+    isField,
     isNumericItem,
     TableCalculation,
 } from '@lightdash/common';
@@ -21,12 +25,8 @@ import {
 } from './SimpleTable.styles';
 
 const SimpleTable: FC = () => {
-    const {
-        resultsData,
-        isLoading,
-        columnOrder: headers,
-        explore,
-    } = useVisualizationContext();
+    const { resultsData, isLoading, columnOrder, explore } =
+        useVisualizationContext();
     const tableItems = resultsData?.rows
         ? getResultValues(resultsData?.rows).slice(0, 25)
         : [];
@@ -44,9 +44,22 @@ const SimpleTable: FC = () => {
         return {};
     }, [explore, resultsData]);
 
-    const rows = mapDataToTable(tableItems, headers);
-    const validData = rows && headers;
+    const rows = mapDataToTable(tableItems, columnOrder);
+    const headers = columnOrder.map((fieldId) => {
+        const field = itemMap && itemMap[fieldId];
+        if (isAdditionalMetric(field)) {
+            // AdditionalMetric
+            return getAdditionalMetricLabel(field);
+        } else if (isField(field)) {
+            // Field
+            return getItemLabel(field);
+        } else {
+            //TableCalculation
+            return friendlyName(fieldId);
+        }
+    });
 
+    const validData = rows && headers;
     if (isLoading) return <LoadingChart />;
 
     return (
@@ -58,7 +71,7 @@ const SimpleTable: FC = () => {
                             <TableHeader>
                                 <tr>
                                     {headers.map((header: string) => (
-                                        <th>{friendlyName(header)}</th>
+                                        <th>{header}</th>
                                     ))}
                                 </tr>
                             </TableHeader>